+++ src/php/Cron.php
@@ -1,6 +1,7 @@
 <?php
 require_once('ApiRequest.php');
 require_once('DatabaseConnection.php');
+require_once('FuelPriceStats.php');
 
 /*
  * This file is part of the fuelPriceStats package.
@@ -36,6 +37,15 @@ class Cron {
      * @return int the return value to be returned from to cron
      */
     public function runJob() {
+        
+        // check config
+        $config = FuelPriceStats::checkConfig($this->config);
+        if($config['result'] == 'ERROR') {
+            echo 'config not valid: '.PHP_EOL.implode(PHP_EOL, $config['messages']);
+            return 1;
+        } else {
+            $this->config = $config['config'];
+        }
         
         // delay api request for random time
         if($this->config['delayRandom'] === true) {