+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - add classes and methods to request tankerkönig api and store values in sqlite database
 - add internal api to retrieve the price data from database as JSON
 - add random delay in execution to fulfill the tankerkoenig.de terms of use
+- add check if config is valid
 
 ### Changed